+++ winpkg/CHANGELOG.md
@@ -5,3 +5,13 @@
 
 - Features
 	- Add/remove install directory to user PATH variable
+
+# v1.1.7
+- General fixes and code improvements
+- Support for multiple command line arguments
+- Update installer output
+- Update uninstaller
+	- Add Publisher, InstallDate, HelpLink fields
+	- Fix environment variable handling issue
+	- Turn echo off in cmd window
+